requirements.txt
fastapi==0.120.0
websockets==15.0.1
python-dotenv==1.1.1
openai==2.6.0
uvicorn[standard]==0.38.0
aiofiles==25.1.0
certifi==2025.10.5